requirements.txt
bcrypt==4.1.2
click==8.1.7
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mysql-connector-python==8.0.28
protobuf==4.25.2
SQLAlchemy==1.4.0
typing_extensions==4.9.0
Werkzeug==2.0.0
WTForms==3.1.2
